# [crypto] replace aes-gcm crate with ring crate in noise.rs