--- a/crates/aptos-crypto/src/noise.rs
+++ b/crates/aptos-crypto/src/noise.rs
@@ -10,6 +10,11 @@
 //! Note that to benefit from hardware support for AES, you must build this crate with the following
 //! flags: `RUSTFLAGS="-Ctarget-cpu=skylake -Ctarget-feature=+aes,+sse2,+sse4.1,+ssse3"`.
 //!
+//! Note that we cannot use the harder-to-shoot-yourself-in-the-foot types of the `ring` crate such
+//! as  OpeningKey/SealingKey and NonceSequence, since Noise's key derivation is non-trivial and
+//! requires us to work with keys as byte arrays. As a result, we make use of `LessSafeKey`,
+//! `UnboundKey` and `Nonce::assume_unique_for_key`.
+//!
 //! Usage example:
 //!
 //! ```
@@ -57,14 +62,11 @@
 //!
 #![allow(clippy::integer_arithmetic)]
 
-use crate::{hash::HashValue, hkdf::Hkdf, traits::Uniform as _, x25519};
-use aes_gcm::{
-    aead::{generic_array::GenericArray, Aead, AeadInPlace, NewAead, Payload},
-    Aes256Gcm,
-};
+use crate::{hash::HashValue, hkdf::Hkdf, traits::Uniform as _, x25519, ValidCryptoMaterial};
+use ring::aead::{self, Aad, LessSafeKey, UnboundKey};
 use sha2::Digest;
 use std::{
-    convert::TryFrom as _,
+    convert::{TryFrom as _, TryInto},
     io::{Cursor, Read as _, Write as _},
 };
 use thiserror::Error;
@@ -116,6 +118,13 @@ pub const fn handshake_resp_msg_len(payload_len: usize) -> usize {
     let enc_payload_len = encrypted_len(payload_len);
     //
     e_len + enc_payload_len
+}
+
+/// Convenience method to wrap an `&[u8]` AES key into a `LessSafeKey` type of the `ring` crate
+fn aes_key(key: &[u8]) -> LessSafeKey {
+    LessSafeKey::new(
+        UnboundKey::new(&aead::AES_256_GCM, key).expect("Unexpected AES256-GCM key length"),
+    )
 }
 
 /// This implementation relies on the fact that the hash function used has a 256-bit output
@@ -302,20 +311,16 @@ impl NoiseConfig {
         let k = mix_key(&mut ck, &dh_output)?;
 
         // -> s
-        let aead = Aes256Gcm::new(GenericArray::from_slice(&k));
+        let aead = aes_key(&k[..]);
+        let mut in_out = self.public_key.to_bytes();
+        let nonce = aead::Nonce::assume_unique_for_key([0u8; AES_NONCE_SIZE]);
 
-        let msg_and_ad = Payload {
-            msg: self.public_key.as_slice(),
-            aad: &h,
-        };
-        let nonce = GenericArray::from_slice(&[0u8; AES_NONCE_SIZE]);
-        let encrypted_static = aead
-            .encrypt(nonce, msg_and_ad)
+        aead.seal_in_place_append_tag(nonce, Aad::from(&h), &mut in_out)
             .map_err(|_| NoiseError::Encrypt)?;
 
-        mix_hash(&mut h, &encrypted_static);
+        mix_hash(&mut h, &in_out[..]);
         response_buffer
-            .write(&encrypted_static)
+            .write(&in_out[..])
             .map_err(|_| NoiseError::ResponseBufferTooSmall)?;
 
         // -> ss
@@ -323,21 +328,17 @@ impl NoiseConfig {
         let k = mix_key(&mut ck, &dh_output)?;
 
         // -> payload
-        let aead = Aes256Gcm::new(GenericArray::from_slice(&k));
+        let aead = aes_key(&k[..]);
+        let mut in_out = payload.unwrap_or(&[]).to_vec();
+        let nonce = aead::Nonce::assume_unique_for_key([0u8; AES_NONCE_SIZE]);
 
-        let msg_and_ad = Payload {
-            msg: payload.unwrap_or(&[]),
-            aad: &h,
-        };
-        let nonce = GenericArray::from_slice(&[0u8; AES_NONCE_SIZE]);
-        let encrypted_payload = aead
-            .encrypt(nonce, msg_and_ad)
+        aead.seal_in_place_append_tag(nonce, Aad::from(&h), &mut in_out)
             .map_err(|_| NoiseError::Encrypt)?;
 
-        mix_hash(&mut h, &encrypted_payload);
+        mix_hash(&mut h, &in_out[..]);
 
         response_buffer
-            .write(&encrypted_payload)
+            .write(&in_out[..])
             .map_err(|_| NoiseError::ResponseBufferTooSmall)?;
 
         // return
@@ -382,17 +383,12 @@ impl NoiseConfig {
 
         // <- payload
         let offset = cursor.position() as usize;
-        let received_encrypted_payload = &cursor.into_inner()[offset..];
 
-        let aead = Aes256Gcm::new(GenericArray::from_slice(&k));
-
-        let nonce = GenericArray::from_slice(&[0u8; AES_NONCE_SIZE]);
-        let ct_and_ad = Payload {
-            msg: received_encrypted_payload,
-            aad: &h,
-        };
-        let received_payload = aead
-            .decrypt(nonce, ct_and_ad)
+        let aead = aes_key(&k[..]);
+        let mut in_out = cursor.into_inner()[offset..].to_vec();
+        let nonce = aead::Nonce::assume_unique_for_key([0u8; AES_NONCE_SIZE]);
+        let plaintext = aead
+            .open_in_place(nonce, Aad::from(&h), &mut in_out)
             .map_err(|_| NoiseError::Decrypt)?;
 
         // split
@@ -400,7 +396,7 @@ impl NoiseConfig {
         let session = NoiseSession::new(k1, k2, rs);
 
         //
-        Ok((received_payload, session))
+        Ok((plaintext.to_vec(), session))
     }
 
     //
@@ -459,18 +455,14 @@ impl NoiseConfig {
             .read_exact(&mut encrypted_remote_static)
             .map_err(|_| NoiseError::MsgTooShort)?;
 
-        let aead = Aes256Gcm::new(GenericArray::from_slice(&k));
-
-        let nonce = GenericArray::from_slice(&[0u8; AES_NONCE_SIZE]);
-        let ct_and_ad = Payload {
-            msg: &encrypted_remote_static,
-            aad: &h,
-        };
-        let rs = aead
-            .decrypt(nonce, ct_and_ad)
+        let aead = aes_key(&k[..]);
+        let mut in_out = encrypted_remote_static.to_vec();
+        let nonce = aead::Nonce::assume_unique_for_key([0u8; AES_NONCE_SIZE]);
+        let rs: &[u8] = aead
+            .open_in_place(nonce, Aad::from(&h), &mut in_out)
             .map_err(|_| NoiseError::Decrypt)?;
-        let rs = x25519::PublicKey::try_from(rs.as_slice())
-            .map_err(|_| NoiseError::WrongPublicKeyReceived)?;
+
+        let rs = x25519::PublicKey::try_from(rs).map_err(|_| NoiseError::WrongPublicKeyReceived)?;
         mix_hash(&mut h, &encrypted_remote_static);
 
         // <- ss
@@ -481,21 +473,17 @@ impl NoiseConfig {
         let offset = cursor.position() as usize;
         let received_encrypted_payload = &cursor.into_inner()[offset..];
 
-        let aead = Aes256Gcm::new(GenericArray::from_slice(&k));
-
-        let nonce = GenericArray::from_slice(&[0u8; AES_NONCE_SIZE]);
-        let ct_and_ad = Payload {
-            msg: received_encrypted_payload,
-            aad: &h,
-        };
+        let aead = aes_key(&k[..]);
+        let mut in_out = received_encrypted_payload.to_vec();
+        let nonce = aead::Nonce::assume_unique_for_key([0u8; AES_NONCE_SIZE]);
         let received_payload = aead
-            .decrypt(nonce, ct_and_ad)
+            .open_in_place(nonce, Aad::from(&h), &mut in_out)
             .map_err(|_| NoiseError::Decrypt)?;
         mix_hash(&mut h, received_encrypted_payload);
 
         // return
         let handshake_state = ResponderHandshakeState { h, ck, rs, re };
-        Ok((rs, handshake_state, received_payload))
+        Ok((rs, handshake_state, received_payload.to_vec()))
     }
 
     /// A responder can respond to an initiator by calling this function with the state obtained,
@@ -544,19 +532,16 @@ impl NoiseConfig {
         let k = mix_key(&mut ck, &dh_output)?;
 
         // -> payload
-        let aead = Aes256Gcm::new(GenericArray::from_slice(&k));
-
-        let msg_and_ad = Payload {
-            msg: payload.unwrap_or(&[]),
-            aad: &h,
-        };
-        let nonce = GenericArray::from_slice(&[0u8; AES_NONCE_SIZE]);
-        let encrypted_payload = aead
-            .encrypt(nonce, msg_and_ad)
+        let aead = aes_key(&k[..]);
+        let mut in_out = payload.unwrap_or(&[]).to_vec();
+        let nonce = aead::Nonce::assume_unique_for_key([0u8; AES_NONCE_SIZE]);
+        aead.seal_in_place_append_tag(nonce, Aad::from(&h), &mut in_out)
             .map_err(|_| NoiseError::Encrypt)?;
-        mix_hash(&mut h, &encrypted_payload);
+
+        mix_hash(&mut h, &in_out[..]);
+
         response_buffer
-            .write(&encrypted_payload)
+            .write(&in_out[..])
             .map_err(|_| NoiseError::ResponseBufferTooSmall)?;
 
         // split
@@ -650,13 +635,17 @@ impl NoiseSession {
         }
 
         // encrypt in place
-        let aead = Aes256Gcm::new(GenericArray::from_slice(&self.write_key));
+        let write_key = aes_key(&self.write_key[..]);
+
         let mut nonce = [0u8; 4].to_vec();
         nonce.extend_from_slice(&self.write_nonce.to_be_bytes());
-        let nonce = GenericArray::from_slice(&nonce);
+        assert_eq!(nonce.len(), 12);
+        let nonce = aead::Nonce::assume_unique_for_key(
+            nonce.try_into().expect("Incorrect AES256-GCM nonce length"),
+        );
 
-        let authentication_tag = aead
-            .encrypt_in_place_detached(nonce, b"", message)
+        let authentication_tag = write_key
+            .seal_in_place_separate_tag(nonce, aead::Aad::empty(), message)
             .map_err(|_| NoiseError::Encrypt)?;
 
         // increment nonce
@@ -666,7 +655,7 @@ impl NoiseSession {
             .ok_or(NoiseError::NonceOverflow)?;
 
         // return a subslice without the authentication tag
-        Ok(authentication_tag.to_vec())
+        Ok(authentication_tag.as_ref().into())
     }
 
     /// decrypts a message from the other peer (post-handshake)
@@ -689,19 +678,23 @@ impl NoiseSession {
         }
 
         // decrypt in place
-        let aead = Aes256Gcm::new(GenericArray::from_slice(&self.read_key));
+        let read_key = aes_key(&self.read_key[..]);
 
         let mut nonce = [0u8; 4].to_vec();
         nonce.extend_from_slice(&self.read_nonce.to_be_bytes());
-        let nonce = GenericArray::from_slice(&nonce);
+        assert_eq!(nonce.len(), 12);
+        let nonce = aead::Nonce::assume_unique_for_key(
+            nonce.try_into().expect("Wrong AES256-GCM nonce length"),
+        );
 
-        let (buffer, authentication_tag) = message.split_at_mut(message.len() - AES_GCM_TAGLEN);
-        let authentication_tag = GenericArray::from_slice(authentication_tag);
-        aead.decrypt_in_place_detached(nonce, b"", buffer, authentication_tag)
+        read_key
+            .open_in_place(nonce, aead::Aad::empty(), message)
             .map_err(|_| {
                 self.valid = false;
                 NoiseError::Decrypt
             })?;
+
+        let (buffer, _authentication_tag) = message.split_at_mut(message.len() - AES_GCM_TAGLEN);
 
         // increment nonce
         self.read_nonce = self