### Description

AES encryption in our custom [Noise protocol](http://www.noiseprotocol.org/) implementation was a bit slow on Apple M1.
This PR switches out the `aes-gcm` crate with the `ring` crate for a faster AES implementation.

We get ~17x improvement on Apple M1 and ~3x improvement on x86.

### Previous M1 performance

```
noise-transport/AES-GCM throughput
                        time:   [59.676 us 59.742 us 59.820 us]
                        thrpt:  [130.60 MiB/s 130.77 MiB/s 130.92 MiB/s]
```

### Better M1 performance

```
noise-transport/AES-GCM throughput
                        time:   [3.2379 us 3.2386 us 3.2393 us]
                        thrpt:  [2.3553 GiB/s 2.3558 GiB/s 2.3562 GiB/s]
                 change:
                        time:   [-94.573% -94.562% -94.551%] (p = 0.00 < 0.05)
                        thrpt:  [+1735.2% +1739.1% +1742.7%]
                        Performance has improved.
```

### Previous x86 performance (from @msmouse)

```
transport/AES-GCM throughput
                        time:   [8.3833 us 8.4206 us 8.4648 us]
                        thrpt:  [922.94 MiB/s 927.78 MiB/s 931.91 MiB/s]
  5 (5.00%) high mild
```

### New x86 performance (from @msmouse)
```
noise-transport/AES-GCM throughput
                        time:   [3.4196 us 3.4330 us 3.4469 us]
                        thrpt:  [2.2134 GiB/s 2.2224 GiB/s 2.2311 GiB/s]
```