@msmouse 2022-06-20
On my lxc inside of Ubuntu 21.04 on dual Xeon 2690v4:

before the change:
```
transport/AES-GCM throughput
                        time:   [8.3833 us 8.4206 us 8.4648 us]
                        thrpt:  [922.94 MiB/s 927.78 MiB/s 931.91 MiB/s]
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild
```

after the change:
```
noise-transport/AES-GCM throughput
                        time:   [3.4196 us 3.4330 us 3.4469 us]
                        thrpt:  [2.2134 GiB/s 2.2224 GiB/s 2.2311 GiB/s]
```